Moved rllib_proxy/_patched/_trainer.py to proxy/agents/trainer.py
