+++ python/ray/rllib/agents/sac/sac/rllib_proxy/_constants.py
@@ -1,7 +1,3 @@
-# By convention, metrics from optimizing the loss can be reported in the
-# `grad_info` dict returned by learn_on_batch() / compute_grads() via this key.
-LEARNER_STATS_KEY = "learner_stats"
-
 # Default policy id for single agent environments
 DEFAULT_POLICY_ID = "default_policy"
