+++ python/ray/rllib/agents/sac/sac/rllib_proxy/_tf_policy_template.py
@@ -1,7 +1,7 @@
-from ray.rllib.agents.sac.sac.rllib_proxy._constants import LEARNER_STATS_KEY
-from ray.rllib.agents.sac.sac.rllib_proxy._utils import add_mixins
 from ray.rllib.agents.sac.sac.rllib_proxy._tf_policy import DynamicTFPolicy
 from ray.rllib.agents.sac.sac.rllib_proxy._needs_patches import TFPolicy
+from ray.rllib.agents.sac.proxy.utils import add_mixins
+from ray.rllib.evaluation.metrics import LEARNER_STATS_KEY
 from ray.rllib.evaluation.policy_graph import PolicyGraph as Policy
 from ray.rllib.utils.annotations import DeveloperAPI, override
