+++ python/ray/rllib/agents/sac/sac/rllib_proxy/_utils.py
@@ -1,21 +1,6 @@
 import tensorflow as tf
 
 
-def add_mixins(base, mixins):
-    """Returns a new class with mixins applied in priority order."""
-
-    mixins = list(mixins or [])
-
-    while mixins:
-
-        class new_base(mixins.pop(), base):
-            pass
-
-        base = new_base
-
-    return base
-
-
 def executing_eagerly():
     try:
         return tf.executing_eagerly()
